requirements.txt
fastapi
uvicorn
jinja2
httpx
pdfminer.six
pydantic
unidecode
python-multipart
